rest of folders moved
All remaining folders were moved
