+++ src/data/models/data-range.model.ts
