+++ src/presentation/controllers/account/index.ts
@@ -1,0 +1,2 @@
+
+export * from './account.controller';
