+++ src/presentation/controllers/customer/index.ts
@@ -1,0 +1,2 @@
+
+export * from './customer.controller';
