+++ src/presentation/controllers/deposit/index.ts
@@ -1,0 +1,2 @@
+
+export * from './deposit.controller';
