+++ src/presentation/controllers/index.ts
@@ -1,1 +1,4 @@
+export * from './account';
+export * from './customer';
+export * from './deposit';
 export * from './security';
