+++ src/presentation/controllers/security/index.ts
@@ -1,1 +1,2 @@
+
 export * from './security.controller';
